backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.28"
psycopg2-binary = "^2.9.9"
supabase = "^2.7.3"
fastapi = {extras = ["standard"], version = "^0.112.1"}
blinker = "1.7.0"
openai = "^1.42.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pydantic-settings = "^2.4.0"
alembic = "^1.14.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
boto3 = "^1.35.68"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"